[reports] Allow to dm the bot for current reports

+++ commands/reports.js
@@ -6,16 +6,16 @@ module.exports = {
   aliases     : [],
   description : 'Get the number of reports you have made',
   args        : [],
-  guildOnly   : true,
+  guildOnly   : false,
   cooldown    : 3,
   botperms    : ['SEND_MESSAGES'],
   userperms   : ['SEND_MESSAGES'],
   execute     : async (msg, args) => {
-    msg.delete().catch(e=>error('Unable to delete message:\n', `\tMessage: ${e.message}\n`, `\tError No: ${e.errno}\n`, `\tCode: ${e.code}\n`));
-    const reports = await getUserReports(msg.member.id);
+    if (msg.channel.type !== 'dm') msg.delete().catch(e=>error('Unable to delete message:\n', `\tMessage: ${e.message}\n`, `\tError No: ${e.errno}\n`, `\tCode: ${e.code}\n`));
+    const reports = await getUserReports(msg.author.id);
     return msg.reply(`You have made ${reports} reports!`).then(m=>{
       setTimeout(()=>{
-        m.delete().catch(e=>error('Unable to delete message:\n', `\tMessage: ${e.message}\n`, `\tError No: ${e.errno}\n`, `\tCode: ${e.code}\n`));
+        if (msg.channel.type !== 'dm') m.delete().catch(e=>error('Unable to delete message:\n', `\tMessage: ${e.message}\n`, `\tError No: ${e.errno}\n`, `\tCode: ${e.code}\n`));
       }, 10000);
     });
   }
